+++ include/multiformats/multibase/multibase.h
@@ -42,29 +42,33 @@ typedef enum
 /**
  * @brief Encode data into a multibase string using the specified encoding.
  *
+ * The output string is null-terminated.
+ *
  * @param base       One of the MULTIBASE_BASE* values (e.g. MULTIBASE_BASE58_BTC).
  * @param data       Pointer to raw binary data to encode.
  * @param data_len   Number of bytes in `data`.
  * @param out        Buffer to write the resulting string (including prefix).
  * @param out_len    Size of `out` in bytes.
- * @return The number of characters written (excluding the terminating null byte) on success,
- *         or a negative value on error.
+ * @return ptrdiff_t
+ *         On success, the number of characters written (excluding the terminating null byte);
+ *         on error, a negative error code (one of multibase_error_t).
  */
-int multibase_encode(multibase_t base, const uint8_t *data, size_t data_len, char *out,
-                     size_t out_len);
+ptrdiff_t multibase_encode(multibase_t base, const uint8_t *data, size_t data_len, char *out, size_t out_len);
 
 /**
- * @brief Decode a multibase string (which includes the prefix) into binary data using the specified
- * encoding.
+ * @brief Decode a multibase string (which includes the prefix) into binary data.
+ *
+ * The input string must be null-terminated.
  *
  * @param base       One of the MULTIBASE_BASE* values indicating the expected encoding.
  * @param in         Null-terminated string with a multibase prefix.
  * @param out        Buffer for decoded bytes.
  * @param out_len    Size of `out` in bytes.
- * @return The number of bytes decoded on success,
- *         or a negative value on error.
+ * @return ptrdiff_t
+ *         On success, the number of bytes decoded;
+ *         on error, a negative error code (one of multibase_error_t).
  */
-int multibase_decode(multibase_t base, const char *in, uint8_t *out, size_t out_len);
+ptrdiff_t multibase_decode(multibase_t base, const char *in, uint8_t *out, size_t out_len);
 
 #ifdef __cplusplus
 }
